add deleteCart method + some fixes

+++ src/main/java/ru/itis/master/party/dormdeals/dto/CartDto.java
@@ -17,5 +17,5 @@ import java.util.List;
 @Data
 public class CartDto {
     private List<ProductDtoCart> productDtoCart;
-    private Integer sumOfProducts;
+    private Double sumOfProducts;
 }
